udpates to the vps backup scripts

+++ vps_swift_backup2.sh
@@ -27,7 +27,7 @@ if which virsh >/dev/null 2>&1; then
 	exit;
  fi
  set -x
- /admin/swift/mkdir_p vps$id --force
+ /admin/swift/mkdir_p vps${id} --force
  sizebytes="$(lvdisplay /dev/vz/${vzid} --units B |grep "LV Size" | awk '{ print $3 }')"
  sizebuffer=10000000000
  lvcreate -L$((${sizebytes} + ${sizebuffer}))B -nimage_storage vz
@@ -37,8 +37,8 @@ if which virsh >/dev/null 2>&1; then
  lvcreate --size 1000m --snapshot --name snap$id /dev/vz/$vzid
  sync
  time qemu-img convert -c -p -O qcow2 /dev/vz/snap${id} /vz/image_storage/${image} 
- /admin/swift/isrm vps$id ${image}
- time /admin/swift/isput vps$id /vz/image_storage/${image} 
+ /admin/swift/isrm vps${id} ${image}
+ time /admin/swift/isput vps${id} /vz/image_storage/${image} 
  umount /vz/image_storage
  echo y | lvremove /dev/vz/image_storage
  rmdir /vz/image_storage
@@ -54,16 +54,28 @@ else
 	curl --connect-timeout 60 --max-time 240 -k -d action=backup_status -d vps_id=${id} "$url" 2>/dev/null
 	exit;
  fi
- /admin/swift/mkdir_p vps$id --force
+ /admin/swift/mkdir_p vps${id} --force
  mkdir -p /vz/${image}
- rsync  -aH --delete -x --no-whole-file --inplace --numeric-ids --exclude=/home/virtfs /vz/root/${vzid}/ /vz/${image}
- vzctl suspend $vzid
- rsync  -aH --delete -x --no-whole-file --inplace --numeric-ids --exclude=/home/virtfs /vz/private/${vzid}/ /vz/${image}
- vzctl resume $vzid
+ if [ -e /vz/private/${id}/root.hdd/root.hdd ]; then 
+  UUID="$(uuidgen)"
+  vzctl snapshot $id --id "$UUID" --skip-suspend --skip-config
+  vzctl snapshot-mount $id --id "$UUID" --target /vz/${image}
+ else
+  rsync  -aH --delete -x --no-whole-file --inplace --numeric-ids --exclude=/home/virtfs /vz/root/${vzid}/ /vz/${image}
+  vzctl suspend $vzid
+  rsync  -aH --delete -x --no-whole-file --inplace --numeric-ids --exclude=/home/virtfs /vz/private/${vzid}/ /vz/${image}
+  vzctl resume $vzid
+ fi
  cd /vz
- /admin/swift/fly vps$id ${image} delete
- /admin/swift/fly vps$id ${image}
- /bin/rm -rf /vz/${image}
+ /admin/swift/fly vps${id} ${image} delete
+ /admin/swift/fly vps${id} ${image}
+ if [ -e /vz/private/${id}/root.hdd/root.hdd ]; then
+  vzctl snapshot-umount $id --id "$UUID"
+  vzctl snapshot-delete $id --id "$UUID"
+  rmdir /vz/${image}
+ else 
+  /bin/rm -rf /vz/${image}
+ fi
 fi
 curl --connect-timeout 60 --max-time 240 -k -d action=backup_status -d vps_id=${id} "$url" 2>/dev/null
